+++ src/ducks/frontEnd.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const url = '/api/';
-const ab = require('./actionBuilders');
+const ab = require('./frontEndABs');
 const {getTrip} = ab;
 
 const G_ICON_CHANGE = 'G_ICON_CHANGE'
