+++ src/ducks/frontEnd.test.js
@@ -1,4 +1,4 @@
-const ab = require('./actionBuilders');
+const ab = require('./frontEndABs.js');
 
 const url = 'http://localhost:3010/api/';
